Add test for bounds and refactor where conversion happens in coil
